requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.1
yfinance>=0.2.40
groq>=0.9.0
deepgram-sdk>=3.2.0
textblob>=0.18.0.post0